Create Adobe-PDF-search-split.vbs
This script requires Adobe Standard and PDF Toolkit to be installed on the machine for it to function. 
If used for automation it requires a user to be logged in as the use of the Adobe SDK that opens the document will actually display the document on the screen. I have not yet found a method to get around this.